requirements.txt
torch
TTS
flask
scipy
llama-cpp-python
flask-cors
